+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added Nginx and PHP-FPM templates to project
+- Added missing current symlink
+
+### Fixed
+- Updated Nginx and PHP-FPM templates to point log folders to deployment
+- Aliases now point to the correct domains by default
 
 
 ## [v1.0.1] - 2020-07-29